icons: stop tracking ~/.icons/
The title is partially misleading, as this also alters .profile to set
the XCURSOR_PATH envvar (sue me--I thought it'd be funny to have two
adjacent commits have completely opposing titles).

Turns out the XCURSOR_PATH envvar fixes the issue with both flameshot
and steam (and is also more elegant since it doesn't need a new
dotfile), and for now that's all of the ones I've had issues with so
this is what I'll go with.