wallpapers: move to ~/.local/share/backgrounds
Also adds a symlink `selected` in the same directory to allow more
easily changing the wallpaper. `feh` (in the qtile autostart script) now
sets `~/.local/share/backgrounds/selected` as the wallpaper, so updating
what file the symlink points to (including files in a different
directory, like a system folder) will update what wallpaper is set
automatically on re-login. Might also be interesting to eventually add
some sort of daemon to detect a change in the symlink and update the
wallpaper immediately.