scripts,mime,xfce,envvars: create dmenu-browser
`~/.local/bin/dmenu-browser` is a script that serves as a launcher for
several browsers (which I will hopefully update as I install and
uninstall browsers). The script works by taking in a URL as the first
argument and then prompting the user for which browser they'd like to
use with dmenu, and then opens that browser with the url.

Changes:
- Creates dmenu-browser script (described above)
- Sets most (hopefully all? but definitely not yet) mime handlers to
  dmenu-browser where appropriate.
  - I felt like it was better to just reset the current mime handlers
    altogether to make this easier, so they were all deleted. I plan on
    setting them back in the future (I have already done so for some of
    the easier mimetypes) but it's really boring so that will take a
    while.
- Sets dmenu-browser as the default web browser in xfce (obviously as a
  custom web browser)
- Sets dmenu-browser as $BROWSER envvar (used to be qutebrowser)