btop: set theme to TTY, "cpu lazy" proc sorting
Although my terminal and even Linux TTY use nord theme, the btop nord
color theme is too monochrome and made it hard to read. I've set it to
TTY so it uses the 16-color pallete (or does it use 8-color?) I've set
for my terminal, which makes it more readable while keeping the soft,
easy-on-the-eyes nord colors.